# Optimize for neon