Neonにはbit reverse命令があるので、find msbをfind lsbに帰着することができ、これは`x & -x`ででき、prefix orよりも若干高速になった

CPU Turing RK1 (Rockchip RK3588, Cortex-A76x4+Cortex-A55x4)
|task|Before|After|
|:---:|:---:|:---:|
|bench_flip|123,624us|118,731us|